requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycryptodome==3.18.0
pyftpdlib==1.5.7
Werkzeug==2.3.6
